+++ templates/partials/controls.php
@@ -56,9 +56,11 @@
 		<span class="artist" title="{{ secondaryTitle() }}">{{ secondaryTitle() }}</span>
 	</div>
 	<div ng-show="currentTrack" class="progress-info">
-		<span ng-show="!loading" class="muted">{{ position.current | playTime }}</span><span
-			ng-show="!loading && durationKnown()" class="muted">/{{ position.total | playTime }}</span>
-		<span ng-show="loading" class="muted">Loading...</span>
+		<div class="progress-text">
+			<span ng-show="!loading" class="muted">{{ position.current | playTime }}</span><span
+				ng-show="!loading && durationKnown()" class="muted">/{{ position.total | playTime }}</span>
+			<span ng-show="loading" class="muted">Loading...</span>
+		</div>
 		<div class="progress">
 			<div class="seek-bar" ng-click="seek($event)" ng-style="{'cursor': seekCursorType}">
 				<div class="buffer-bar" ng-style="{'width': position.bufferPercent, 'cursor': seekCursorType}"></div>